refactor: output scp to templates
