+++ packages/preset-umi/src/constants.ts
@@ -4,4 +4,3 @@ export const DEFAULT_PORT = '8000';
 export const DEFAULT_HOST = '0.0.0.0';
 export const CACHE_DIR_NAME = '.cache';
 export const TEMPLATES_DIR = join(__dirname, '../templates');
-export const PREFETCH_ROUTE_MAP_SCP_TYPE = 'umi-route-chunk-files-map';
