+++ packages/preset-umi/templates/routePrefetch/prefetchRouteFilesScp.js
